python/requirements.txt
google-cloud-vision==3.7.2
